scraper/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.2.1
psycopg-binary==3.2.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
